requirements.txt
fuse-python
kubernetes
